+++ src/app/article/index.js
@@ -1,7 +1,8 @@
 import {memo, useCallback, useMemo} from 'react';
 import {useParams} from "react-router-dom";
-import useStore from "../../store/use-store";
-import useSelector from "../../store/use-selector";
+import useStore from "../../hooks/use-store";
+import useSelector from "../../hooks/use-selector";
+import useTranslate from "../../hooks/use-translate";
 import useInit from "../../hooks/use-init";
 import PageLayout from "../../components/page-layout";
 import Head from "../../components/head";
@@ -9,12 +10,10 @@ import Navigation from "../../containers/navigation";
 import Spinner from "../../components/spinner";
 import ArticleCard from "../../components/article-card";
 import LocaleSelect from "../../containers/locale-select";
-import useTranslate from "../../hooks/use-translate";
 
 function Article() {
 
   const store = useStore();
-
   const params = useParams();
 
   useInit(() => {
