+++ src/app/basket/index.js
@@ -1,11 +1,12 @@
 import {memo, useCallback} from 'react';
-import useStore from "../../store/use-store";
-import useSelector from "../../store/use-selector";
+import useStore from "../../hooks/use-store";
+import useSelector from "../../hooks/use-selector";
+import useInit from "../../hooks/use-init";
+import useTranslate from "../../hooks/use-translate";
 import ItemBasket from "../../components/item-basket";
 import List from "../../components/list";
 import ModalLayout from "../../components/modal-layout";
 import BasketTotal from "../../components/basket-total";
-import useTranslate from "../../hooks/use-translate";
 
 function Basket() {
 
@@ -17,14 +18,14 @@ function Basket() {
     sum: state.basket.sum
   }));
 
-  const {t} = useTranslate();
-
   const callbacks = {
     // Удаление из корзины
     removeFromBasket: useCallback(_id => store.actions.basket.removeFromBasket(_id), [store]),
     // Закрытие любой модалки
     closeModal: useCallback(() => store.actions.modals.close(), [store]),
   }
+
+  const {t} = useTranslate();
 
   const renders = {
     itemBasket: useCallback((item) => (
