+++ src/app/index.js
@@ -1,9 +1,8 @@
 import {useCallback, useContext, useEffect, useState} from 'react';
 import {Routes, Route} from 'react-router-dom';
+import useSelector from "../hooks/use-selector";
 import Main from "./main";
 import Basket from "./basket";
-import useStore from "../store/use-store";
-import useSelector from "../store/use-selector";
 import Article from "./article";
 
 /**
