+++ src/app/main/index.js
@@ -1,5 +1,6 @@
 import {memo} from 'react';
-import useStore from "../../store/use-store";
+import useStore from "../../hooks/use-store";
+import useTranslate from "../../hooks/use-translate";
 import useInit from "../../hooks/use-init";
 import Navigation from "../../containers/navigation";
 import PageLayout from "../../components/page-layout";
@@ -7,7 +8,6 @@ import Head from "../../components/head";
 import CatalogFilter from "../../containers/catalog-filter";
 import CatalogList from "../../containers/catalog-list";
 import LocaleSelect from "../../containers/locale-select";
-import useTranslate from "../../hooks/use-translate";
 
 function Main() {
