+++ src/containers/locale-select/index.js
@@ -1,8 +1,8 @@
 import {memo, useCallback, useMemo} from "react";
-import useStore from "../../store/use-store";
-import useSelector from "../../store/use-selector";
-import Select from "../../components/select";
+import useStore from "../../hooks/use-store";
+import useSelector from "../../hooks/use-selector";
 import useTranslate from "../../hooks/use-translate";
+import Select from "../../components/select";
 
 function LocaleSelect() {
