+++ src/hooks/use-store.js
@@ -1,5 +1,5 @@
 import {useContext} from "react";
-import {StoreContext} from "./context";
+import {StoreContext} from "../store/context";
 
 /**
  * Хук для доступа к объекту хранилища
